modified:   src/main/java/com/freesia/metatradepublisher/model/ItemInfo.java 	modified:   src/main/java/com/freesia/metatradepublisher/model/Trade.java 	modified:   src/main/java/com/freesia/metatradepublisher/rpc/proto/fake-trade.proto

+++ src/main/java/com/freesia/metatradepublisher/model/ItemInfo.java
@@ -1,5 +1,5 @@
 package com.freesia.metatradepublisher.model;
 
-public record ItemInfo(String id, double amount, String store_address, String description) {
+public record ItemInfo(String id, long amount, String store_address, String description) {
     
 }
