+++ src/main/java/com/freesia/metatradepublisher/model/Trade.java
@@ -6,14 +6,14 @@ import com.freesia.metatradepublisher.CryptoUtils;
 public class Trade {
     private String senderAddress;
     private String receiverAddress;
-    private double amount;
-    private double commission;
+    private long amount;
+    private long commission;
     private long timestamp;
     private String signature;
     private String senderPublicKey;
     private String description;
 
-    public Trade(String senderAddress, String receiverAddress, double amount, double commission, long timestamp, String signature, String senderPublicKey, String description) {
+    public Trade(String senderAddress, String receiverAddress, long amount, long commission, long timestamp, String signature, String senderPublicKey, String description) {
         this.senderAddress = senderAddress;
         this.receiverAddress = receiverAddress;
         this.amount = amount;
@@ -50,25 +50,25 @@ public class Trade {
 
 
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
 
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
 
 
 
-    public double getCommission() {
+    public long getCommission() {
         return commission;
     }
 
 
 
-    public void setCommission(double commission) {
+    public void setCommission(long commission) {
         this.commission = commission;
     }
